split run_24kHz
split a long file into 3 components; main_run, config, prep_24kHz
